Working on some refactoring
