+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Router, Redirect } from '@reach/router';
+import { Router, Redirect, navigate } from '@reach/router';
 import ChooseYourContact from './components/ChooseYourContact';
 import { firebase, db } from './firebase';
-import Login from './components/Login';
+import Login from './components/Auth/Login';
 import DashBoard from './components/DashBoard';
 import ScheduleFreeCall from './components/ScheduleFreeCall';
 import SchedulePaidCall from './components/SchedulePaidCall';
@@ -11,6 +11,8 @@ import CallConfirmation from './components/CallConfirmation';
 import AboutUs from './components/AboutUs';
 import PreviousCalls from './components/dashboard/PreviousCalls';
 import NavBar from './components/NavBar';
+import SignUp from './components/Auth/SignUp';
+import MobileLandingPage from './components/MobileLandingPage';
 
 import UpdateAccount from './components/UpdateAccount';
 import { fetchUser } from './app/utils';
@@ -44,27 +46,32 @@ function useAuth() {
 }
 
 function App() {
-  const user = useAuth();
-  console.log(user, 'USERRRRR');
-  return user ? (
+  const result = firebase.auth().onAuthStateChanged(user => {
+    if (user) {
+      navigate(`/user/${user.uid}`);
+    } else {
+      navigate('/');
+    }
+  });
+
+  return (
     <>
-      <NavBar user={user} />
+      {/* <NavBar user={user} /> */}
       <Router>
-        <Redirect from='/' to={`user/${user.uid}`} noThrow />
-        <DashBoard user={user} path='/user/:userId' />
-        <ChooseYourContact user={user} path='/choose/:userId' />
+        {/* <Redirect from='/' to={`user/${user.uid}`} noThrow /> */}
+        <DashBoard user={result.user} path='/user/:userId' />
+        <ChooseYourContact user={result.user} path='/choose/:userId' />
         <ChooseCallPlan path='/choose/:userId/:contactId/call-plan' />
         <ScheduleFreeCall path='/choose/:userId/:contactId/:frequency/schedule-free' />
         <SchedulePaidCall path='/choose/:userId/:contactId/:frequency/schedule' />
         <CallConfirmation path='/confirmation/:contactId' />
         <PreviousCalls path='prev-calls/:userId' />
         <AboutUs path='/about-us' />
-        <UpdateAccount user={user} path='/account/:userId' />
+        <UpdateAccount user={result.user} path='/account/:userId' />
+        <Login path='/login' />
+        <SignUp path='/signup' />
+        <MobileLandingPage path='/' exact />
       </Router>
-    </>
-  ) : (
-    <>
-      <Login />
     </>
   );
 }
